requirements.txt
anyio==3.6.1
autopep8==1.7.0
fastapi==0.85.0
greenlet==1.1.3
idna==3.4
install==1.3.5
lxml==4.9.1
numpy==1.23.3
pandas==1.5.0
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pydantic==1.10.2
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.2.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
toml==0.10.2
typing_extensions==4.3.0
